refactor: move auth module to the root of the folder components
