+++ src/components/auth.ts
@@ -1,6 +1,6 @@
 import './auth.css';
-import { Api } from "../../api/api";
-import { CreateUserBody, SingInRequestBody } from "../../api/typeApi";
+import { Api } from "../api/api";
+import { CreateUserBody, SingInRequestBody } from "../api/typeApi";
 export async function registration() {
   removePopUp();
   const backBlack = document.createElement('div');
